added divide + bute->float

+++ Features/SerializableColor.cs
@@ -9,7 +9,7 @@ namespace WarheadCustomColor.Features
         {
             
         }
-        public SerializableColor(byte r, byte g, byte b, byte w)
+        public SerializableColor(float r, float g, float b, float w)
         {
             R = r;
             G = g;
@@ -17,26 +17,26 @@ namespace WarheadCustomColor.Features
             W = w;
         }
 
-        public SerializableColor(Vector4 color): this((byte)color.x, (byte)color.y, (byte)color.z, (byte)color.w){}
+        public SerializableColor(Vector4 color): this((float)color.x, (float)color.y, (float)color.z, (float)color.w){}
 
         /// <summary>
         /// Red color
         /// </summary>
-        public byte R { get; set; } = 0;
+        public float R { get; set; } = 0;
         /// <summary>
         /// Green color
         /// </summary>
-        public byte G { get; set; } = 0;
+        public float G { get; set; } = 0;
 
         /// <summary>
         /// Blue color
         /// </summary>
-        public byte B { get; set; } = 0;
+        public float B { get; set; } = 0;
         
         /// <summary>
         /// Brightness color
         /// </summary>
-        public byte W { get; set; } = 0;
+        public float W { get; set; } = 0;
         
         public static implicit operator Color(SerializableColor a) => new Color(a.R, a.G, a.B);
         public static implicit operator Vector4(SerializableColor a) => new Vector4(a.R, a.G, a.B, a.W);
